+++ day07.hpp
@@ -6,54 +6,54 @@ namespace example
 {
     namespace N1
     {
-        static constexpr Int result = 43210;
-        static constexpr std::initializer_list<Int> code{3,  15, 3,  16, 1002, 16, 10, 16, 1,
-                                                         16, 15, 15, 4,  15,   99, 0,  0};
-        static constexpr std::array<Int, 5> sequence{4, 3, 2, 1, 0};
+        static constexpr long long result = 43210;
+        static constexpr std::initializer_list<long long> code{3,  15, 3,  16, 1002, 16, 10, 16, 1,
+                                                               16, 15, 15, 4,  15,   99, 0,  0};
+        static constexpr std::array<long long, 5> sequence{4, 3, 2, 1, 0};
     }
 
     namespace N2
     {
-        static constexpr Int result = 54321;
-        static constexpr std::initializer_list<Int> code{3,  23, 3,  24,  1002, 24, 10, 24, 1002,
-                                                         23, -1, 23, 101, 5,    23, 23, 1,  24,
-                                                         23, 23, 4,  23,  99,   0,  0};
-        static constexpr std::array<Int, 5> sequence{0, 1, 2, 3, 4};
+        static constexpr long long result = 54321;
+        static constexpr std::initializer_list<long long> code{
+            3, 23, 3,  24, 1002, 24, 10, 24, 1002, 23, -1, 23, 101,
+            5, 23, 23, 1,  24,   23, 23, 4,  23,   99, 0,  0};
+        static constexpr std::array<long long, 5> sequence{0, 1, 2, 3, 4};
     }
 
     namespace N3
     {
-        static constexpr Int result = 65210;
-        static constexpr std::initializer_list<Int> code{
+        static constexpr long long result = 65210;
+        static constexpr std::initializer_list<long long> code{
             3,  31, 3,  32, 1002, 32, 10, 32, 1001, 31, -2, 31, 1007, 31, 0, 33, 1002,
             33, 7,  33, 1,  33,   31, 31, 1,  32,   31, 31, 4,  31,   99, 0, 0,  0};
-        static constexpr std::array<Int, 5> sequence{1, 0, 4, 3, 2};
+        static constexpr std::array<long long, 5> sequence{1, 0, 4, 3, 2};
     }
 
     namespace N4
     {
-        static constexpr Int result = 139629729;
-        static constexpr std::initializer_list<Int> code{3,  26, 1001, 26, -4, 26, 3, 27, 1002, 27,
-                                                         2,  27, 1,    27, 26, 27, 4, 27, 1001, 28,
-                                                         -1, 28, 1005, 28, 6,  99, 0, 0,  5};
-        static constexpr std::array<Int, 5> sequence{9, 8, 7, 6, 5};
+        static constexpr long long result = 139629729;
+        static constexpr std::initializer_list<long long> code{
+            3,  26, 1001, 26,   -4, 26, 3,  27,   1002, 27, 2,  27, 1, 27, 26,
+            27, 4,  27,   1001, 28, -1, 28, 1005, 28,   6,  99, 0,  0, 5};
+        static constexpr std::array<long long, 5> sequence{9, 8, 7, 6, 5};
     }
 
     namespace N5
     {
-        static constexpr Int result = 18216;
-        static constexpr std::initializer_list<Int> code{
+        static constexpr long long result = 18216;
+        static constexpr std::initializer_list<long long> code{
             3,    52,   1001, 52,   -5,   52, 3,  53, 1,    52, 56, 54, 1007, 54, 5,
             55,   1005, 55,   26,   1001, 54, -5, 54, 1105, 1,  12, 1,  53,   54, 53,
             1008, 54,   0,    55,   1001, 55, 1,  55, 2,    53, 55, 53, 4,    53, 1001,
             56,   -1,   56,   1005, 56,   6,  99, 0,  0,    0,  0,  10};
-        static constexpr std::array<Int, 5> sequence{9, 7, 8, 5, 6};
+        static constexpr std::array<long long, 5> sequence{9, 7, 8, 5, 6};
     }
 }
 
 namespace input
 {
-    static constexpr std::initializer_list<Int> data{
+    static constexpr std::initializer_list<long long> data{
         3,    8,    1001, 8,    10,   8,     105,  1,    0,    0,    21,   34,  43,   64,   85,
         98,   179,  260,  341,  422,  99999, 3,    9,    1001, 9,    3,    9,   102,  3,    9,
         9,    4,    9,    99,   3,    9,     102,  5,    9,    9,    4,    9,   99,   3,    9,
