+++ day09.hpp
@@ -3,7 +3,7 @@
 
 namespace input
 {
-    static constexpr std::initializer_list<Int> data{
+    static constexpr std::initializer_list<long long> data{
         1102, 34463338, 34463338, 63,    1007,  63,    34463338, 63,    1005,  63,    53,    1102,
         1,    3,        1000,     109,   988,   209,   12,       9,     1000,  209,   6,     209,
         3,    203,      0,        1008,  1000,  1,     63,       1005,  63,    65,    1008,  1000,
