+++ day17.hpp
@@ -33,7 +33,7 @@ namespace example
 
 namespace input
 {
-    static constexpr std::initializer_list<Int> data{
+    static constexpr std::initializer_list<long long> data{
         1,     330,   331,   332,   109,   3762,  1101,  1182,  0,     16,    1101,  0,     1467,
         24,    101,   0,     0,     570,   1006,  570,   36,    101,   0,     571,   0,     1001,
         570,   -1,    570,   1001,  24,    1,     24,    1105,  1,     18,    1008,  571,   0,
